Update LazyDFU for Minecraft 22w15a/1.19
Mojang's DFU library can now skip "optimizing" rules and there is a native abstraction in the Minecraft codebase. This makes the mod much simpler, since all we need to do now is just overwrite enableDataFixerOptimizations with a no-op method.

It's still an unknown if Mojang will implement this mod in the base game, but until then... it's still here.

My last parting gift before an indeterminate period of absence from Minecraft because of my move to NYC!